Refactor recipes endpoints into router
Moved the solitary recipes endpoint to routers/recipes.py, thus setting
the stage for a more scalable app. I want to get most of the structure
of the app nailed down early on so I don't need to copy and paste large
chunks of code later on and deal with rewiring imports.