Add database operation functions to create and read from database
I found a better workaround for the pyarango bugs I describe in the dev
journal: buried within the instance of SimpleQuery is the response data
is a plain ol' list. I have to do this because the toJson() method for
query results isn't working properly. I plan to submit a PR to pyArango
at some point with a fix for this bug.